search rendering correctly
